Added a very basic controller class.
I would consider using an actual MVC system, but the UI isn't the interesting part of this project. This will be fine for now.
