+++ Dragon Warrior Monsters 2 Randomizer/dwm2rando.css
@@ -10,3 +10,11 @@ form > .container-fluid > .row{
 form .row b {
 	color: #F33;
 }
+
+.error_message {
+	border: 2px solid red;
+	background-color: #FDD;
+	width: 100%;
+	margin: 5px;
+	padding: 5px;
+}
